polish split
